Initial Commit!
This is an example react/redux app.  It was built as a proof of concept under an informal contract that was never paid in full.  Rather than waste time trying to litigate for the remaining , I've decided to just upload this as a portfolio piece.  It's been scrubbed of any potential copyright or trademarked terms.